rpm,deb: package mgr-orchestrator in separated package
ceph-mgr depends on it in the sense that ceph-mgr-orchestrator is
a hard dependency of ceph-mgr. like cmake depends on cmake-data,
even if the later cannot function on its own without the former.

the reason why we need a separated package is that
ceph-mgr-orchestrator is written in python, and should be arch
independent. it helps the downstream to understand this, and to have
less disk footprint when it's populated as prebuilt packages.

Signed-off-by: Kefu Chai <kchai@redhat.com>