mgr/cephadm: do not refresh device inventory on mgr restart
The service inventory is more fluid and is faster to gather.  We also
make a time-saving assumption that we don't need to persist our cache
updates when making changes because we know a mgr restart will refresh.

Device inventory changes are much less frequent and slower.  Let's not
refresh them every restart.

Signed-off-by: Sage Weil <sage@redhat.com>