mon/Monitor: fail forwarded tell commands
If we have a mix of pre-octopus and octopus+ mons, the old mon may forward
a tell command.  If we try to process it via the asok/tell queue, we will
assert out when we try to send a reply message (see AnonConnection
impl).  Instead, just reply with an error, and a hint that the command
is failing becuase the cluster is mid-upgrade.

Fixes: https://tracker.ceph.com/issues/44297
Signed-off-by: Sage Weil <sage@redhat.com>