mgr/DaemonServer: ensure pg state is fully up to date before merging pgs
It's possible to have slightly old PG state that says both PGs are
active+clean and can be merged when in fact they cannot.

One way this can happen is if one merge participant has upmap and one
does not.  On one interation, adjust_pgs() will clear the upmap.  On the
next check 2 seconds later, the upmap is gone and the PGs still appear
as active+clean because the pg state is not fresh.

Fixes: https://tracker.ceph.com/issues/44684
Signed-off-by: Sage Weil <sage@redhat.com>