cephadm: monitoring: fix Prometheus "root fs full" alert
which would never be triggered due to the file system being mounted on
"/rootfs" (instead of "/") and also exported as such in the metric
labels.

Fixes metrics exported that aren't where they would expected to be.
After fix, they appear as they would on a node-exporter instance running
directly on a host.

Fixes: https://tracker.ceph.com/issues/44781

Signed-off-by: Patrick Seidensal <pseidensal@suse.com>