mon/MgrMap: use always_on_modules from most recent release
If our release isn't in the always_on_map, say because you just upgraded,
then start with the modules from the last/most recent release.  This
should be sufficient to allow an upgrade when the mon is still the old
release and the mgrmap doesn't tell us what modules we should be enabling.

(Arguably, the always_on_modules should be enshrined in the MgrMap itself
and not in the mon?)

Signed-off-by: Sage Weil <sage@redhat.com>