mds: introduce cephfs.mirror.info virtual xattr
This virtual extended attribute is to be used when setting up
a filesystem for mirroring. The idea is to prevent multiple
ceph clusters adding a same remote ceph cluster as a peer for
a filesystem.

Adding a peer is configured via an interface exposed by manager
module ("mirroring" plugin). This extended attribute is set on
the remote filesystems root directory by the manager plugin
thereby preventing other clusters adding the filesystem as a
peer.

Note that this extended attribute is constraint to be set only
on the root directory. Technically, filesystems in multiple
ceph clusters can mirror distinct directories to a filesystem
in the remote cluster. When supporting this, the path constraint
can be removed for setting the extended attribute on directories.

Signed-off-by: Venky Shankar <vshankar@redhat.com>