cephadm: auto wrap and unwrap ipv6 addresses
This patch attempts to simplify IPv6 support in cephadm by automatically
wrapping and unwrapping IPv6 addresses when required.

There are some asumptions though, if you are supplyings an IPv6 addrv
then it needs to be wrapped. But because you are specifiying, you should
know what your doing.

But in general, it means in bootstrap you should be able to supply ipv6
addresses wrapped or not so long as there isn't a post appended.

Fixes: https://tracker.ceph.com/issues/46922
Signed-off-by: Matthew Oliver <moliver@suse.com>