rbd-nbd: make list-mapped work for restarted daemons
If a daemon in "persist" mode is restarted, /sys/block/nbdX/pid is
set to a garbadge.

To find the rbd-nbd process in this case we have to parse ps
output looking for a command line with `--device /dev/nbdX` arg,
which should always be present for daemons started in "persist"
mode.

Signed-off-by: Mykola Golub <mgolub@suse.com>